requirements.txt
asgiref==3.4.1
Django==3.2.18
django-cors-headers==3.10.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2023.2
sqlparse==0.4.3
typing_extensions==4.1.1
